starboard: Fix unsafe integer in AudioTrack sink
Fix potential overflow and narrowing issues in AudioTrackAudioSink.
Change accumulated_written_frames, last_playback_head_position,
and playback_head_position to int64_t to prevent overflow after
long playback times (~12.4 hours at 48kHz).
Update GetFramesDurationUs to accept int64_t.

Bug: 511356295
TAG=agy
CONV=0d3a836a-08df-4860-a867-0f6affc85b25

+++ starboard/android/shared/audio_track_audio_sink_type.h
@@ -169,7 +169,7 @@ class AudioTrackAudioSink : public SbAudioSinkImpl {
 
   void ReportError(bool capability_changed, const std::string& error_message);
 
-  int64_t GetFramesDurationUs(int frames) const;
+  int64_t GetFramesDurationUs(int64_t frames) const;
 
   const raw_ptr<Type> type_;
   const int channels_;
